+++ src/subcommands/runtime/select.ts
@@ -5,8 +5,8 @@ import { compareVersions } from "../../compareVersions.js";
 import { addCreateClientOptions, createClient } from "../../createClient.js";
 import { addLogLevelOptions, createLogger } from "../../logLevel.js";
 import { UserInputError } from "../../types/UserInputError.js";
-import { fallbackAlias } from "./aliasGeneration.js";
-import { resolveLatestAlias, resolveUniqueAlias } from "./aliasResolution.js";
+import { fallbackAlias } from "./helpers/aliasGeneration.js";
+import { resolveLatestAlias, resolveUniqueAlias } from "./helpers/aliasResolution.js";
 
 async function selectRuntimeEngine(
   logger: SimpleLogger,
